requirements.txt
django==3.2.6
djangorestframework==3.12.4
django-filter==2.4.0
dj-database-url==0.5.0
gunicorn==20.1.0
django-cors-headers==3.7.0
sendgrid==6.8.0
razorpay==1.2.0
mysqlclient==2.0.3
pillow==8.3.2
django-import-export==2.5.0
oauth2client==4.1.3
google-auth==1.34.0
google-api-python-client==2.15.0
pandas
django-prometheus==2.1.0
jinja2==3.0.1
whitenoise
